requirements.txt
# API & web
fastapi>=0.115.2,<1.0
uvicorn[standard]==0.30.6
httpx==0.27.2

# LLM
google-genai>=1.0.0

# Embeddings
sentence-transformers==3.1.1

# Vector store
chromadb==0.5.15

# PDF extraction
pymupdf==1.24.11
pdfplumber==0.11.4

# Data
pandas==2.2.3
openpyxl==3.1.5
pydantic==2.9.2
pydantic-settings==2.5.2

# Config
python-dotenv==1.0.1
pyyaml==6.0.2

# UI
gradio==6.12.0

# Utils
tqdm==4.66.5
tiktoken>=0.7.0
rapidfuzz==3.10.0
numpy==1.26.4
